api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==44.0.1
dnspython==2.8.0
fastapi==0.115.0
frozenlist==1.7.0
google-auth==2.41.1
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.6.4
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.6.1
pydantic_core==2.33.2
pyfcm==2.0.7
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.3
rsa==4.9.1
sentry-sdk==2.16.0
sniffio==1.3.1
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
